requirements.txt
fastapi==0.115.11
alembic==1.15.1
SQLAlchemy==2.0.39
sqladmin==0.20.1
bcrypt-4.3.0 
passlib-1.7.4
aiosqlite-0.21.0
pymongo-4.12.0
motor-3.7.0